moodev bank implement

+++ db/migrate/20181002023050_add_money_to_wallet.rb
@@ -1,5 +1,5 @@
 class AddMoneyToWallet < ActiveRecord::Migration[5.0]
   def change
-    add_column :wallets, :money, :integer
+    add_column :wallets, :money, :decimal
   end
 end
